app/pyproject.toml
[tool.poetry]
name = "currencytest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
lxml = "^4.9.1"
bs4 = "^0.0.1"
sqlalchemy = "^1.4.44"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
pyjwt = "^2.6.0"
py-bcrypt = "^0.4"


[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

